Added basic support for .mtl files.
The .mtl files can be parsed, subject to a few (heavy) limitations and
the model displayed in colour.

The limitations of the parser are:
- The path to a .mtl file is relative to your current directory, rather
  than the .obj file which references it. So unless you launched the
  objviewer from the directory containing your .obj file, it won't be
  able to load the mtl file(s).
- The illum directive for materials is ignored (it's always assumed to
  be 2).
- The texture map directives (map_Ka, map_Kd, etc.) are ignored.

The renderer is also limited at the moment:
- It will ignore all material info except the Kd (diffuse colour)
  attribute.
- It doesn't support lighting yet; everything is flat shaded.

I added one other new feature to the renderer: you can toggle between
lines and polygons by pressing the 'p' key. Cool huh?