New camera representation.
The camera is now represented as a reference point (the point that it's looking
at, a.k.a the target), a rotation angle around each of the X, Y and Z axes and a
distance from the target.

This actually makes it a lot simpler, although it took a bit of fiddling to get
right. It also means the camera controls feel a bit more natural.