Limiting the maximum size of my display lists, for a massive performance improvement.
Introduced the RenderGroup class, which associates a material with range
of display list IDs. This lets me have more than one display list per
material and therefore that I can limit the number of faces per display
list (which was the root of the problem before).

Rendering is now just a case of iterating over the list of render groups
and calling all of their display lists in turn.

As a side effect of the way I create the render groups, I'm not doing
any extra processing for materials that are declared, but not used in
the model - which probably helps as well.

I've currently got the maximum number of faces per render group set at
20000. On my MacBook Pro with an NVIDIA 8600M GPU, this gives roughly a
3x speedup.