Got rid of Frames from the Model.
Frames were my original abortive attempt at loading keyframed
animations. There are much better ways to do it and, since animation
playback isn't implemented yet anyway, it was just getting in the way.

So I got rid of it and will put in some replacement when I put in the
animation support.