Got basic transparency working.
Added support for the materials dissolve attribute and texture and the
Renderer::prepare method now ensures that materials with any transparency
are drawn after any opaque materials.

It's still not quite right, but it's an improvement. I think the way to get
it working properly from here is to adjust the texture combining parameters
(see the glTexEnvi calls in loadTexture).