Got VBOs working (with a few bugs).
The speed is extremely variable. FPS changes a lot even while you're not
doing anything. It's always better than the display-list based version
though - in some cases 5 times faster!

There's a segfault when rendering a model without a material. I've also
managed to disable the wireframe mode (again!) because it simply doesn't
know how to render like that. But it's still progress. :-)