Rendering fixes for models with no material and materials with cut-out textures.
The mapD textures are handled correctly now, which is good. The FairyForest model
looks much nicer. You can also specify whether a texture is a matte when loading it
to the GPU and it'll be handled correctly.

Models without explicit normals are handled better now, and I've fixed a few bugs
related to normals.

Also, to save client-side memory, I delete the pixel from images after uploading
it to the GPU.