WIP commit of the new animation system. Still has lots of bugs.
Introduced a Curve class, to hold keyframe data and interpolate between
keyframes.

All vertices, texture coords, normals and colours in the model are
stored as curves now, instead of plain old Float4 objects. Same goes for
the data in RenderGroups.

Changed the rendering process so that each RenderGroup streams data for
the current animation frame into the vertex buffer whenever the current
frame time changes.

Changed the model loading process. It's now the OBJViewerApp class that
implements the ParserCallbacks interface. It loads multiple model files,
treating each new file as a new set of keyframe data for the existing
model. It assumes that the number of vertices and the faces don't
change from frame to frame, but that their positions, normals, texture
coordinates and colours may do.

Known bugs at this stage include:
- The face data is being parsed or rendered incorrectly.
- There seem to be some points popping up in the wrong positions.
- Normal calculations are a bit off.
- Textures aren't displaying.
- Gets extremely slow when several frames of animation are loaded.