Moved to having 1 shader program per render group, instead of one global program.
There's now two different shader programs that we use: one for objects
with a material, which takes parameters specifying which textures are
active; and another for objects without a material.

Also added some code to count the number of animated curves, as a
precursor to only storing & updating the bits that are animated.