Extended the math library with vectors and matrices of other sizes.
Removed math3d.(h|cpp) and replaced it with vector.(h|cpp). We now have
Vector2, Vector3 and Vector4 classes instead of the old Float4 classes
and the code has been updated to use Vector4. There are also both
Matrix3 and Matrix4 classes, where before there was only Matrix4. There
are a bunch of new operations on each of the types, too.

Finally, the new types are all in a namespace, vh:: (which should help
avoid confusion).