Big refactoring to use my new VGL library.
Not all features work as well as before yet, but it can load and display
the .obj files and .ply files I tried it on.

Missing features include:
- Camera movement is not yet as good as before.
- Quick camera positioning (keys 0-6) are no longer working.
- Good initial camera position not being calculated.
- There seems to be an awful lot of z-fighting going on.
- Per-vertex colours aren't working any more.

Note that although vector.cpp is no longer used in the app, I haven't
deleted it yet because I want to make sure I've grabbed all the useful
code in it for VGL's equivalent classes.

Still to do:
- Refactor VGL's Camera interface to make it easier to plug your own
  implementation in. Then, make OBJViewer's camera class override it and
  use that instead.
- Fix the issues above.