Added a basic & hacky CMakeLists.txt to build the objviewer.
I'm not too happy with the way this works yet, but it does work at least.

The problem is that it should only depend on the output of the VGL build -
ideally by doing an installation into the OBJViewer build directory, then
referencing the headers and shared libs from there.

What it does at the moment is hardcode the include dir for VGL headers to be the
VGL source directory (double-plus-ungood) and the link dir for the VGL library
to be a certain location inside the OBJViewer build dir (not as bad, but still
not ideal). Plus it relies on hackiness to make it work: according to CMake
docs, you're not supposed to be able to refer to a subdirectory in an
add_dependencies() call. Sigh.